backend/requirements.txt
django==4.2
djangorestframework==3.14
django-cors-headers==3.13.0

python-decouple==3.8
python-slugify==8.0.4
djangorestframework-simplejwt==5.3.1
django-filter==24.3
djangorestframework-recursive==0.1.2
drf-yasg

pillow
pytesseract

mysqlclient